requirements.txt
flatten_dict==0.4.2
pytest==8.0.0
ruamel.base==1.0.0
